Extract reverse addon
